// src/generated/prisma/enums.ts
export const Role = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  PRESIDENT: 'PRESIDENT',
  SECRETARY: 'SECRETARY',
  DIRECTOR: 'DIRECTOR',
  BOARD_MEMBER: 'BOARD_MEMBER',
  VIEWER: 'VIEWER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const EventStatus = {
  DRAFT: 'DRAFT',
  SUBMITTED: 'SUBMITTED',
  UNDER_REVIEW: 'UNDER_REVIEW',
  CORRECTION_REQUIRED: 'CORRECTION_REQUIRED',
  APPROVED: 'APPROVED',
  ARCHIVED: 'ARCHIVED'
} as const

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus]


export const EventType = {
  PHYSICAL: 'PHYSICAL',
  ONLINE: 'ONLINE',
  HYBRID: 'HYBRID'
} as const

export type EventType = (typeof EventType)[keyof typeof EventType]


export const OrgType = {
  ROTARY_CLUB: 'ROTARY_CLUB',
  ROTARACT_CLUB: 'ROTARACT_CLUB',
  NGO: 'NGO',
  SCHOOL: 'SCHOOL',
  COLLEGE: 'COLLEGE',
  UNIVERSITY: 'UNIVERSITY',
  CORPORATE: 'CORPORATE',
  GOVERNMENT: 'GOVERNMENT',
  COMMUNITY_ORG: 'COMMUNITY_ORG',
  OTHER: 'OTHER'
} as const

export type OrgType = (typeof OrgType)[keyof typeof OrgType]


export const FundingSource = {
  CLUB_FUND: 'CLUB_FUND',
  ROTARY_CLUB: 'ROTARY_CLUB',
  SPONSOR: 'SPONSOR',
  DONATION: 'DONATION',
  MEMBER_CONTRIBUTION: 'MEMBER_CONTRIBUTION',
  PARTNER_ORGANIZATION: 'PARTNER_ORGANIZATION',
  OTHER: 'OTHER'
} as const

export type FundingSource = (typeof FundingSource)[keyof typeof FundingSource]


export const BeneficiaryCategory = {
  CHILDREN: 'CHILDREN',
  STUDENTS: 'STUDENTS',
  WOMEN: 'WOMEN',
  MEN: 'MEN',
  SENIOR_CITIZENS: 'SENIOR_CITIZENS',
  PATIENTS: 'PATIENTS',
  FAMILIES: 'FAMILIES',
  SCHOOL_STUDENTS: 'SCHOOL_STUDENTS',
  COLLEGE_STUDENTS: 'COLLEGE_STUDENTS',
  DIFFERENTLY_ABLED: 'DIFFERENTLY_ABLED',
  GENERAL_PUBLIC: 'GENERAL_PUBLIC',
  COMMUNITY: 'COMMUNITY',
  ENVIRONMENT: 'ENVIRONMENT',
  ANIMALS: 'ANIMALS',
  ROTARACTORS: 'ROTARACTORS',
  OTHER: 'OTHER'
} as const

export type BeneficiaryCategory = (typeof BeneficiaryCategory)[keyof typeof BeneficiaryCategory]


export const ParticipantGroup = {
  ROTARACTOR: 'ROTARACTOR',
  ROTARIAN: 'ROTARIAN',
  COUNCIL: 'COUNCIL',
  GUEST: 'GUEST'
} as const

export type ParticipantGroup = (typeof ParticipantGroup)[keyof typeof ParticipantGroup]


export const DocumentCategory = {
  POSTER: 'POSTER',
  ATTENDANCE_SHEET: 'ATTENDANCE_SHEET',
  BILL: 'BILL',
  INVOICE: 'INVOICE',
  PERMISSION_LETTER: 'PERMISSION_LETTER',
  APPRECIATION_LETTER: 'APPRECIATION_LETTER',
  CERTIFICATE: 'CERTIFICATE',
  NEWSPAPER_COVERAGE: 'NEWSPAPER_COVERAGE',
  SOCIAL_MEDIA_SCREENSHOT: 'SOCIAL_MEDIA_SCREENSHOT',
  GENERATED_REPORT: 'GENERATED_REPORT',
  OTHER: 'OTHER'
} as const

export type DocumentCategory = (typeof DocumentCategory)[keyof typeof DocumentCategory]


export const SocialPlatform = {
  INSTAGRAM: 'INSTAGRAM',
  FACEBOOK: 'FACEBOOK',
  LINKEDIN: 'LINKEDIN',
  YOUTUBE: 'YOUTUBE',
  WEBSITE: 'WEBSITE',
  OTHER: 'OTHER'
} as const

export type SocialPlatform = (typeof SocialPlatform)[keyof typeof SocialPlatform]